refactor: handleLeave 메서드를 manager로 위임
- 섬에 따라 떠날 때 로직이 달라질 예정이라 위치 변경
