+++ src/domain/components/islands/interface/island-manager.ts
@@ -5,5 +5,6 @@ export interface IslandManager {
     join(player: Player): Promise<void>;
     getActiveUsers(islandId: string, myPlayerId: string): Promise<Player[]>;
     left(islandId: string, playerId: string): Promise<void>;
+    handleLeave(player: Player): Promise<Player>;
     removeEmpty(islandId: string): Promise<void>;
 }
