+++ src/modules/islands/deserted-island-manager.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { DesertedIslandManager } from 'src/domain/components/islands/deserted-storage/deserted-island-manager';
+import { RedisTransactionManagerModule } from 'src/infrastructure/redis/redis-transaction-manger.module';
 import { IslandJoinComponentModule } from 'src/modules/island-joins/island-join-component.module';
 import { DesertedIslandStorageComponentModule } from 'src/modules/islands/deserted-island-storage-component.module';
 import { IslandComponentModule } from 'src/modules/islands/island-component.module';
@@ -11,6 +12,7 @@ import { PlayerStorageComponentModule } from 'src/modules/users/player-storage-c
         PlayerStorageComponentModule,
         IslandJoinComponentModule,
         IslandComponentModule,
+        RedisTransactionManagerModule,
     ],
     providers: [DesertedIslandManager],
     exports: [DesertedIslandManager],
