+++ src/modules/islands/normal-island-manager.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { NormalIslandManager } from 'src/domain/components/islands/normal-storage/normal-island-manager';
+import { RedisTransactionManagerModule } from 'src/infrastructure/redis/redis-transaction-manger.module';
 import { IslandJoinComponentModule } from 'src/modules/island-joins/island-join-component.module';
 import { IslandComponentModule } from 'src/modules/islands/island-component.module';
 import { NormalIslandStorageComponentModule } from 'src/modules/islands/normal-island-storage-component.module';
@@ -11,6 +12,7 @@ import { PlayerStorageComponentModule } from 'src/modules/users/player-storage-c
         PlayerStorageComponentModule,
         IslandJoinComponentModule,
         IslandComponentModule,
+        RedisTransactionManagerModule,
     ],
     providers: [NormalIslandManager],
     exports: [NormalIslandManager],
